chore: rename in resource descendants
